Remove databrowser-example dependency from app
The dependency is not required at the moment, because all Web
Component imports are defined as file reference in
web-component.import.ts. This simplifies initial development.

As soon as stable Web Component are available on an npm registry,
they should be included as as dependency (registration of Web
Components could still be done in the web-component.import.ts file).