Add fix for VSCode Volar plugin
The Volar plugin (https://github.com/johnsoncodehk/volar) supports
Vue3 and is the new recommended VSCode plugin for Vue3 apps.

Unfortunately, the plugin doesn't recognize that axios is part of the
Vue/nuxt application. The type definition file provided by this
commit fixes that issue.

Newer versions of the Volar plugin may provide a solution out of
the box, so it may be a good idea to check if the fix is still needed
from time to time.