Remove test file for generic-pagination
All tests in the file were commented out, so it came to an exception
when the tests were run. This commiit removes the test file. Although
at first it may seem to be a bad decision, one must know that this will
be the last commit for the databrowser application in its current form.
There is an agreement that the databrowser application will be
bootstrapped from 0 with nuxt 3 as dependency. That's the reason why
it doesn't make sense to put additional efforts to fix the tests of this
file.