Switch from Vue WebComponents to pure Vue components
Using Vue WebComponents as done until now in this project requires the
WebComponent template to be compiled on the client, because the
WebComponent template was written as a JavaScript template string.

After internal discussion by the devs we came to the conclusion that it
doesn't make sense to provide WebComponents in the current form at
this project stage.

Therefor, in this commit all previous Vue WebComponents were moved
and transformed to pure Vue components.