Add skipLibCheck = true to tsconfig.json
The setting skipLibCheck skips the type checking of declaration files
(see https://www.typescriptlang.org/tsconfig#skipLibCheck). This
setting was necessary because of a bug in the current build of
vue-json-pretty (see https://github.com/leezng/vue-json-pretty/issues/121).

Please remove the skipLibCheck setting as soon as the bug in the lib
is fixed!