Use scrollbar-gutter CSS style to prevent content jumping
The page content usually jumps when a user navigates from a page with
scrollbar to one without and vice versa.

This project already tried to fix that issue using a margin-left fix,
but that fix prevented the header from having full width.

The new fix uses the scrollbar-gutter CSS style which was created right
for this use case. One drawback is browser support: Chrome supports it
out of the box, but Firefox support will be available only later on
(presumably with version 97). Safary has no support at all as of yet, but
given that Safari uses overlay-scrollbars this seems to be no problem.