Implement POC for in-memory data modification
This first step demonstrates how to handle updates that are emitted by a
component. There is no persistence, many cases won't work yet and some
code refeactoring is necessary, but the principles are established and
should work from here on.

This commit also introduces a diff-view to get an impression on what
data changed. You can find it in the edit view under the "JSON diff" tab.