Implement select box with predefined options
Passing the select options is a bit problematic, because in essence one
wants to pass an array of options, each option having a "value" and a
"label". But that form of data passing is not possible using HTML
attributes, because HTML attributes allow string values only.

The implemented solution is to have a list of attributes named
"value_XXX" and "label_XXX" for each option, e.g.
"value_001" with value "EC" and "label_001" with value "Eurac" are merged
into the select option {value: "EC", label: "Eurac"}.