Reimplement array data behavior together with image gallery edit
Until now, array data behavior was hidden be, hind a list of properties
and it was quite complicated to get it right, especially for the edit
mode.

This commit changes the array data behavior, together with a major
rewrite of the edit image gallery component.

The rewrite of the edit image gallery component was also necessary
to improve its reusability.