Fix bug in delete confirmation dialog
Before this commit, all delete confirmation dialogs used the same pinia
store to control their visibility. Using many such dialogs on the same
page led to a bug, where always the last dialog would be shown. The
reason is, that only the last dialog's visibility was set by the shared
store.

This commit removes the shared store and moves the dialog visibility
handling to the parent component. That way, the visibility is handled per
component, as one would expect.