Update vue, vue-tsc, @vite-js/plugin-vue and fix import declaration errors
This update moves vue, vue-tsc and @vite-js/plugin to its current
version.

Due to the update, a lot of "import declaration errors" emerged, because
the compile macros defineProps, defineEmits, ... don't need to be
imported in the first place. This commit removes all of those macros and
fixes the import declaration errors.