Remove GIT pre-commit from .githooks (keep husky)
The GIT pre-commit hook was removed from the .githooks directory
because it is not needed anymore. The pre-commit hook was used to
invoke the linter before a commit. This is done by husky. The
pre-commit hook was just a leftover from the time before husky was
used.

The advantage of husky is, that it is not necessary to install the
pre-commit hook manually. It is installed automatically when the
project is cloned and npm is installed.