Fix bug when generating Data Browser URLs out of MetaData
This fix is a crude hack. Unfortunately, at the moment  there is no other signal we could use to find out
which API/domain is targeted. It involves hard-coding domain strings.

A better solution would be to have a domain property in the dataset metadata, because that way we can support other domains without code changes.