Remove Quick View and related facilities
The Quick View was intended to provide a nice overview over datasets.

Implemented for the ODHActivityPoi datasets, it turned out that the Quick View it is not needed
at the moment (see issue #552 on GitHub).

Therefor, this commit removes all of the Quick View related code and configs and moves / renames
the shared stuff that is still used to other locations.