Update eslint and associated packages, together with typescript configs
The changes in this commit were necessary due to updates in eslint and its packages. The changes
are aligned with the way new vue projects are configured using "create-vue". This change should
simplify future updates.

Note that a lot of new linting errors arose due to eslint update. The appearing errors are temporarily
disabled (see eslint.config.js) to not overload this commit with file fixes. The errors will be fixed in
the next commits.