Improve map view performance by rendering OverviewCardItem elements only if map is not visible
The map view is implemented as a Dialog over the OverviewListPage. Therefor, all elements of
OverviewListPage are rendered when the map view is visible, although they are not visible to
the user.

This commit adds a check for map view visibility and renders the OverviewListPage content only
if the map is not visible. The result is improved performance when the map view is visible,
especially when navigating from other parts of the Data Browser to the map view.