Remove IconParser and Replace all its usages with fixed icons
The IconParser component, that handles dynamic icons, was implemented for the
QuickView some time ago.

As it is, the QuickView got removed afterwards, the IconParser is not strictly necessary anymore.
It is even a problematic component, because during build time it provokes lots of warnings.

Therefor, this commit removes the IconParser.